requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.2
pytesseract==0.3.4
pytz==2020.1
sqlparse==0.3.1
